Bug#34852090: Incorrect result with VALUES in a correlated LATERAL subquery Bug#35087820: VALUES Statement with dependent subquery is wrong
When a table value constructor has to be read multiple times during a
query, it will only return data the first time it is read. When it's
read the second time, it will return EOF immediately.

It is caused by a bug in TableValueConstructorIterator::Read() where
it returns EOF immediately when the number of examined rows is equal
to the number of rows in the table value constructor. Since this
counter is intentionally not reset when the iterator is
re-initialized, the table value constructor will keep returning EOF
forever once every row has been read, and not start again from the
beginning when it's re-initialized.

Fixed by checking the position of the iterator over the row value list
instead of the number of examined rows.

Change-Id: I0e828eb0de0360a16cea9f77b578c68205cefbaf