Bug#35710218: mysql server failed at Item_singlerow_subselect::val_str
The problem here occurs when matching the result of a row subquery
with a value list in an IN predicate. The function in_row::find_item()
stores the row subquery into appropriate buffers, then skips the test
if any item in the row subquery is NULL. It does this by calling
Item_subselect::is_null(), which calls Item::update_null_value(),
which calls Item::evaluate(). However this function is properly
implemented only for scalar values, not for row values. For
Item_singlerow_subselect, the function attempts to call val_str(),
which looks at Item_singlerow_subselect::m_value. But this is nullptr
for a row subquery, thus the program execution fails.

The fix here is to realize that the function store_value() aggregates
the NULL values of the members of the row subquery, meaning that
Item_singlerow_subselect::null_value is already calculated.
Thus there is no reason to call is_null() and the function call can be
replaced with a check on null_value.

Change-Id: I8d1ce349a848323f8a31f831b40c726d1d0a3c78