Bug#35813134: Optimize table value constructors in the hypergraph optimizer
Run table value constructors through the hypergraph optimizer instead
of using the hard coded path for them in the old optimizer.

This fixes bug#31949057 (TABLE VALUE CONSTRUCTOR FOLLOWED BY ORDER BY
IS NOT SORTED CORRECTLY) in the hypergraph optimizer. The bug is still
present in the old optimizer.

The patch changes JOIN::optimize() so that it passes table value
constructor query blocks to FindBestQueryPlan() in the hypergraph
optimizer. Such query blocks go through the fast path for single-table
queries. The fast path was updated so that it also accepted table-less
queries.

When adding a SORT access path for ORDER BY, we now also check if the
query is table-less, and add a STREAM access path before the sorting
if it is. This is needed because Filesort requires a record buffer,
and we need a temporary table for that when there are no other tables.

Change-Id: Ia1a0ed5ce6711c23e7a3825f602cdb3aa75478b8