Bug#35790087 Enforce programmatic control of dependency tracking
Problem:
The --binlog-transaction-dependency-tracking variable is deprecated and
might be removed from future versions of MySQL Server.

Analysis:
In NDB all binlog content is processed by a single thread so CPU
consumption is a potential bottleneck. In the case where MTA is not
used, calculating writeset hashes is wasted work, and thus it is still
important to be able to turn off the writeset hash calculations on an
individual MySQL Server when transaction dependency calculations are not
necessary.

Fix:
Make it clear that being able to programmatically decide whether
calculation of writeset hashes for transaction dependencies are important
for MySQL NDB Cluster use case by:
- Adding new member variable to injector::transaction that controls
wheter to calculate writeset hashes, improve comments and describe the
use case.
- Improving existing testcase for --ndb-log-dependency-tracking=0 to
fail hard if the setting is not off when test starts.

Change-Id: Idb34df0c554890f9329e52d38f6256847dd6c7a2