Bug#35621842: Rollup subquery segv
This is a simplified patch for 8.0.

The problem here is that a nameless field is presented to the function
find_item_in_list(). Since the search is name-based, it will fail.

The culprit is in Item::split_sum_func2(), which when presented with
an expression that contains a GROUPING field is supposed to split
the function's arguments, however it adds a reference to the expression.

The function is enhanced so that it also checks for this case.

Change-Id: Ib272698e2718c4a65eb8a71a8022d6ac527723e8