Bug#35590108: Assertion `static_cast<size_t>(bit_num) < x.capacity()' failed in IsBitSet()
The hypergraph optimizer hit an assertion when creating ROR Union
plans. This happened if an inexact ROR Union scan could be used for a
query, and the number of predicates in the query was 64 or higher.

The assertion was hit because a bitset representing the predicates
subsumed by the ROR Union scan had too small capacity. Fixed by
allocating a bitset with capacity equal to the number of predicates.

Change-Id: I4420fcfcf6757d90dd79bf9da964c246c9e5a323