Bug#35788971: Unexpected Result by NO_BNL
Wrong results were seen in some queries with outer joins which
contained join conditions that always evaluated to false.

The problem was a bug in the code that checked if a join condition
would reject all rows. It incorrectly assumed that a condition that
referred to a column in a table which was known to be empty or contain
NULL-complemented rows only, would reject all rows. This assumption
does not hold if the condition is not NULL-rejecting on the table (for
example "col IS NOT NULL").

The bug had made it into both of the optimizers. In the old optimizer,
it only affected hash joins, so adding a NO_BNL hint to disable hash
joins would make the query return correct results. In the hypergraph
optimizer, it affected both hash joins and nested loop joins.

It is fixed by making the check stricter, and only prune the join tree
if the condition is NULL-rejecting on the table that is known to be
empty or contain NULL-complemented rows only. The tables on which the
condition is NULL-rejecting are given by Item::not_null_tables().

Change-Id: I554f9411a9256ba8e5aae65c9c84d27a88fd404c