Bug#35710183: mysql server 8.1.0 crashed at Item_ident::walk
While trying to clone a Item_ref object during condition
pushdown, Item_ref::print does not print the underlying
expression since it is an alias. However, in this case
the underlying expression is a user variable, which
results in a mismatch for the number of fields present
in the original expression and the cloned expression
(parser creates a field object for the alias in the
cloned expression while the original expression has
none).
In case of condition pushdown, we always want the
original expression to be printed instead of the alias.
So the solution is to skip the code path which has
special handling for aliases. Also, removed a
(!const_item()) check which was added for a similar
case but for a constant expression.

Change-Id: I3b390915dbc7618175fd1b65e06af73eb0a9f6ef