WL#14595: Refactor GR recovery to do not use View_change_log_event
The VCLE was added to the binary log to avoid introducing stress on the Group
Communication Service (GCS).
Since then GCS was extended with message fragmentation and compression, which
streamlines the delivery of bigger messages.
group_replication_view_change_uuid option is deprecated, when all group members
are version 8.3.0 or higher the option value is ignored.
The GCS will be used to send the information, which was earlier sent through the
VCLE of the binary log.

Change-Id: I7bba1a31ecd3b8a6872f49040935fb2bb7bcf3a3