Bug#35710179: [backport] Mysql server 8.1.0 failed at Item_sum_avg::val_decimal
[ backport: skips added test for hypergraph, which gives wrong result. Fixed on trunk
            with non-regression patch
	    5681b52 Bug#35836231: i_main.subquery-bug35710179 fails with hypergraph
	    This can't be backported anyway due to other trunk changes, in addition to
	    non-regression status
]

The problem is missing propagation of item properties for the left-hand
expression of an IN subquery predicate, which causes a window function
to loose its property as a window function, and there is subsequent
failure because the code expects a window function later on.

The fix is to add that propagation.

Change-Id: Iae71684c38727cc6db182b1873ff561c07e35f9d