Bug#35714466 Timeout too long on MGMAPI connections converted to transporters
When a TransporterRegistry (TR) instance is connecting to an MGMD, it
first connects using MGMAPI, and then converts the connection to a
Transporter connection for further communication.

The initial MGMAPI connection should have a timeout set so that the TR
instance does not spend too long waiting for unsuccessful connection
attempts to time out.

The MGMAPI connection timeout is set equal to the MGMAPI connection timeout
used by the TR for the connection it uses for getting/setting dynamic ports.

Change-Id: I3e3361dff3fca88ca0792c07222629c48359daf8