Bug#35789589 Assertion `m_index_cursor.is_positioned()' failed.
This error can happen for a query where the query plan has:
* A left hash-join H1. The first row it produces must have a
  NULL-row right half.
* Another hash join H2 with H1 as its probe input. H2 must start with
  reading from the probe input (i.e. H1).
* A third hash-join H3. H3 must have H2 as its build input.

With such a plan, H3 calls H2.Init() which causes the first (partial
null) row to be read from H1. H3 then calls H2.SetNullRow(false),
which erroneously resets the null flag in the H1 row.

The fix is to change HashJoinIterator so that it does not call
SetNullRowFlag(false) in this state. In addition, an assert has been
added to HashJoinIterator::SetNullRowFlag() that will fail if the
function is called in that state (i.e. between Init() and the first
call to Read()).

Change-Id: I239fdcb34183529e9e2a6eb3db093f5c5320280d