Bug#35814434 memory-usage increases with connection-sharing and benchmarking
When benchmarking mysqlrouter with connection-sharing enabled,
the RES of the mysqlrouter process increases over time even though it
stays more-or-less constant when running the connection-sharing
disabled.

ASAN reports no memory leaks.

Using "valgring --tool=massif", it is seen that the "timer_ops" are
allocated a lot and not destroyed fast enough. Those timer_ops are
from Timer.async_wait() and are already cancel()ed, but their result
is not yet delivered to the async-callers.

Cancelled timers are expected to be handled with the highest priority,
but here they are handled only if there is no other work.

Change
======

- treat "expired timer" and "cancelled timer" as "execute now".

Change-Id: I76868b52850f81cbcc3f189e6128e4ff98063299