Bug#35627798: MySQL crashes by a createTable-createTrigger-Select sequence
Some queries using window functions could make the server exit.

The resolver does not reserve space in Query_block::base_ref_items for
all items in the ORDER BY and PARTITION BY clauses of window
functions. It reserves space for the fields referenced in those
clauses, but it reserves no space for expressions that don't reference
any fields. It is usually not a problem, because the resolver is
already quite generous with the space it allocated for base_ref_items,
so there is often enough room for these clauses. But it's not
guaranteed.

Fixed by making Item::do_itemize() also count items inside ORDER BY
clauses so that they are accounted for when allocating the
base_ref_items array.

The original logic already double-counted some items, making the array
bigger than it needs to be in most cases. This patch makes it
double-count items in even more cases, so it may be useful to rethink
how the base_ref_items array is allocated in a future patch. For
example, base_ref_items could be changed to a data structure that
could grow dynamically, so that we don't need to figure out up front
how much space is required. The approach in the current patch,
however, is chosen in order to be conservative and safe.

Change-Id: I5527dbb5d141f5d702f74b9b29045eda98568262