Bug#35502793 INFORMATION_SCHEMA.INNODB_FT_DELETED's behavior is different from the manual.
When running OPTIMIZE TABLE on a table with FULLTEXT index, the index
was not fully optimized at the expected time.

A single pass (execution) of OPTIMIZE TABLE by design limits the
number of words that will be optimized. This is specified using the
innodb_ft_num_word_optimize variable, which defaults to 2000,
INNODB_FT_DELETED is only purged when OPTIMIZE has been run enough
times that all words in the table/index have been optimized. However
the bug caused it to need an additional pass over which should
normally be required. This is fixed so that the pass which optimizes
the last batch of words is the one which completes the index
optimization.

Change-Id: I80db913cc3d72035de26ff3107532992bdf79e9a