BUG#35068461 mysql crashes after detecting DEADLOCK
Related:
Bug 109861 mysql crashes after detecting DEADLOCK

Problem
=======

Under load, the server can report a deadlock in innodb,
and exit.

Root cause
==========

Several conditions are required.

1)

The server is under heavy load.

2)

Some applications queries do acquire a lot of data locks in innodb.

3)

One monitoring query executes:
  SELECT * FROM performance_schema.data_locks;

4)

When executing a scan in innodb,
to report rows to populate the performance_schema.data_locks table,
a lot of rows can be reported at once,
causing a lot of memory consumption.

The number of rows reported at once depends on how many locks can
be taken by a single session, see 2)

5)

Because a possibly unbounded number of rows are returned,
the server may allocate un unbounded amounts of memory,
eventually leading to an out of memory exception.

6)

In the implementation for table performance_schema.data_locks,
a C++ try-catch block handles the exception, and cause the SELECT
to fail with an error instead.

7)

From this point, the server is in an incoherent state,
and will soon fail detecting a deadlock.

The root cause is that the try-catch block on 6) can happen
inside a mutex lock - mutex unlock critical section inside innodb,
leaving the critical section with a mutex still on hold.

Fix
===

After careful analysis, it is determined that the try-catch code
was a bad idea, and causes more issues that it attempts to solve,
because the server code in general is not prepared to handle
C++ exceptions.

This try-catch is removed.

Should out of memory conditions still occur, the server will fail
immediately, instead of failing anyway with a deadlock,
so this removal is not a regression.

Note that, independently of this fix,
the memory consumption for table performance_schema.data_locks
has been reduced, which makes the failure less likely.

See related:
Bug 35240825 performance_schema.data_locks query increased RAM usage to OOM

Change-Id: I216390f8777a57f4ad4c490d95aca8cd9d1526e3