Bug#35790381 Assertion `false && "Inconsistent row counts for different AccessPath objects."'
If we aggregate on fields, EstimateAggregateRows() will use indexes or
histograms on those fields for making row estimates. But the code for
doing so lacked the logic for mapping from an Item_ref to an
Item_field (given that this is indeed what the Item_ref points to). So
for Item_ref, EstimateAggregateRows() would ignore indexes and
histograms, and fall back to rules of thumb.

This could again cause inconsistent row estimates, if the same
aggregate term appears as an Item_field in one plan and Item_ref in
another.

The fix for this is to call real_item() for Item_ref aggregate terms,
so that we find the underlying Item_field if there is one. In addition
to eliminating the assertion error, this also means that we get better
row estimates for some queries.

Change-Id: I92816d6e4f9e5dc8daaa6d75780508838cbb2ff2