Bug#35821812 Build breaks on MacOS 13.5.1 with system protobuf for telemetry plugin
For "system" protobuf, we use the cmake function FIND_PACKAGE.
The corresponding cmake module FindProtobuf.cmake may do
  set_property(TARGET protobuf::libprotobuf APPEND PROPERTY
               INTERFACE_COMPILE_FEATURES cxx_std_11
              )
This property will be applied transitively to anything linking
with libprotobuf::libprotobuf, and this may break the build
since we use C++20 features.

The fix is to use ext::libprotobuf throughout the cmake codebase.
We define this target for both "bundled" and "system" protobuf.

Change-Id: I222ef0a4e24a8cc40860a64298003286c68f3ec9