Bug#35767961 NdbSocket uses blocking SSL_shutdown, also for non-blocking sockets
For non blocking sockets using NdbSocket and TLS a switch to blocking
mode was done for shutdown and close.

This could block other threads doing read or write using same NdbSocket
object, or block the thread the connects and disconnects all
transporters.

The switch to blocking mode for non blocking sockets is removed.
And also calling SSL_shutdown is avoided such in the case that shutdown
is first called SSL_shutdown will not be called a second time and
potentially block when close is called.

Change-Id: I95cbe8e98effc32b082dce519c49ff396809c770