Bug#32969309 WAITING FOR SCHEMA MDL FOR SQL THREAD ON SQL NODE
Problem:

On a setup where a server in a replica cluster has set read_only=ON,
when a DROP DATABASE statement is used in the source cluster, the SQL
thread on the replica server is stuck at "Waiting for schema metadata
lock".

Analysis:

At the performance_schema.metadata_locks table in the replica server,
multiple IX MDL locks are granted for each individual schema and one X
is pending. The IX locks were granted to the NDB DD client, while the
pending X lock is at the replica worker applying the DROP DATABASE
statement.

But observing the logs, the server states that it failed to acquire
the IX MDL lock on the schema for the NDB DD client, which is
contraditory. The fact is that the MDL was granted and the read-only
option is what caused the erroneous log message. From here, it can be
observed that the NDB DD client locks were never released because the
read-only check made the function return too early, before registering
these for later release. Since IX MDL cannot satisfy a X MDL request,
then the caller of X MDL would wait indefinitely.

Solution:

Make the NDB DD client code register the MDL request tickets right
after acquiring the locks. Doing this for all functions so that locks
are always registered for release when they are no longer required.

Change-Id: I9cdf423998ef0bf0b4ff778526a97e2181135dee