Bug#35785032 Hypergraph: Wrong first row cost for rowid intersect
For ROWID_INTERSECT access paths, init_cost was set equal to 'cost'.
This is incorrect, as the cost of reading rows from a ROWID_INTERSECT
is approximately proportional to the number of rows read. Therefore,
this commit sets init_cost to the sum of the init_cost of each of the
child paths. (In the common case where the child paths are index range
scans on a non-derived table, init_cost becomes zero.)

Likewise, init_once_cost of ROWID_INTERSECT access paths was set to
zero. This would be wrong if any of the child paths had a non-zero
init_once cost. This patch therefore ensures that init_once_cost is
set to the sum of init_once_cost of each of the child paths.

Change-Id: I701238824f9ce7cf3c8f2f85a14d832892879afe