Bug#35836231: i_main.subquery-bug35710179 fails with hypergraph
Some queries with IN subqueries or quantified comparisons in the
SELECT list returned wrong results with the hypergraph optimizer.

Wrong results were returned because materialized expressions in the
left operand of these predicates were not replaced by references to
the corresponding field in the temporary table holding the
materialized result. This was supposed to happen during the call
FinalizePlanForQueryBlock(), but didn't happen because the helper
function WalkAndReplace() didn't include code for walking into IN or
quantified comparisons.

Fixed by making WalkAndReplace() walk into these expressions
too. (Some repeated code in that function was extracted to a separate
helper function to avoid adding more repetitions of the same code.)

Change-Id: I77a69b3f3d566e759e4b88522f44b62195e39f21