Bug#35768766 - Big numbers loosing precision in client-side applications
Description
===========
As an MRS user, I expect to be able to retrieve records with fixed-point
"DECIMAL"/"NUMERIC" columns, "BIGINT SIGNED" and "BIGINT UNSIGNED" columns
and obtain the exact value in a client-side application.

However, these type of columns are capable of storing values that will
lose precision when converted to the corresponding numeric type ("number")
on a browser-based application written in JavaScript/TypeScript.
This is because the "number" type in JavaScript is only capable of
representing integers in the range -253 + 1 to 253 - 1 (IEEE 754), inclusive.

When building the response to an HTTP request in the scope of MRS,
the MySQL router creates a JSON representation of the record,
where the values of DECIMAL, NUMERIC and BIGINT columns are converted to the
corresponding numeric type in JSON. This is fine, from a JSON standpoint,
however, the native JavaScript JSON parser works in a way that automatically
converts those values into a local "number" type, and thus, they are prone
to losing precision.

Fix
===
MRS plugin now interprets "Accept" http header, to implement
"Server-driven content negotiation". For the REST requests,
the plugin allows following values:

   * application/json
   * application/x.ieee754.client+json,

Where the first one, points that the client may process standard
generated json documents (as it was until now). The second
points that client is not able to interpret "bigint", thus
those values should be serialized as "json-string".

Example
=======
To receive "big-ints" encoded as string, client shoud include
following header into the request:

    Accept: application/x.ieee754.client+json

Including both MIME types, allows the plugin to generate the
response as "application/json".