Bug#35845129: Missing subqueries in EXPLAIN with hypergraph optimizer
When using the hypergraph optimizer, EXPLAIN didn't show the plans for
subqueries in the SELECT list of the query if the result of the
subquery was materialized.

The subquery plans were missing because EXPLAIN looked at a projection
list that contained the temporary table columns instead of the
original expressions, so the subqueries were not seen. With the old
optimizer, this was compensated for by an extra loop through all
temporary tables in the plan and adding the plans for those subqueries
that were not seen in the projection list. This loop did nothing when
using the hypergraph optimizer, since it looked at QEP_TAB objects,
which are created by the old optimizer only.

Fixed by looping through Query_block::base_ref_items instead of
JOIN::fields, since the former contains the original expressions
instead of the temporary table fields for these subqueries.

This made the loop over the QEP_TAB objects redundant, even in the old
optimizer, so the patch also removes that loop.

Change-Id: Ic02b424f022f94a5d0492f8f7a5318ef3ad85ecd