Bug#35846435 Need to pre-pack send buffers for ssl_write
Root cause to the problem where ssl_write() may get a smaller
send buffer on retries is that consolidate() calculated how many *full*
iovec buffers we could fit into buff[MaxTlsRecord].

Thus if the last iovec[] buffer size had increased on a write retry,
consolidate would calculate a smaller buff[] size on the retry
if the last (now larger) iovec[] buffer could not fit anymore.

That was likely based on a misconseption that iovec_data_sent(), which is
used to report the buffer size consumed by 'writev()', only handled
the consumption of a number of full iovec[] buffers. However, that is
not the case: any number of bytes may be reported as consumed, and the
iovec[]' will be correctly updated accordingly.

Patch allowes the full size of the buff[MaxTlsRecord] to be used for
packing, with a partial copy from the last iovec[] in order to completely
fill buff[]. Thus the case where we could no longer fit data from
the last iovec[] buffer is avoided.

Change-Id: I4d0d7d124421d510ab80e0a2566f6565785ad6d3