Bug#35813111: Use C++ standard library for more bit operations [my_bit_log2, noclose]
Remove the my_bit_log2() function from my_bit.h. It is only used on
the form 1U << my_bit_log2(x), which produces the same result as
std::bit_floor(), except when x is zero. Since it is only used for
adjusting the MyISAM block size value to a power of two, and the block
size already has to be in the range [1024, 16384], the difference on
zero input is irrelevant.

Change-Id: I5c5aee86fd4e562f2c3e89fcd4ac3822646ac5d7