Bug#35667611 Data node failure during the rolling restart, halts the restart process
The management server maintains a cache mapping nodeids to host address
presentation format strings (e.g. 33 -> "47.184.12.55")

These cached strings are returned to MGM protocol requests for information
about connected nodes.

The cache is currently :
 - Populated when an MGM protocol request is made for a destination
   which is connected at the transporter level, and has no cached
   address string
 - Cleared when processing a NODE_FAILREP or NF_COMPLETEREP signal

This is problematic when the disconnection of a remote node is slow, as the
following sequence is possible :
 1  Data node 5's address is cached 'host A'
 2  Data node 5 vanishes (network loss)
 3  Data node 5's failure is detected by MGMD and a disconnect is requested
 4  Other data nodes detect node 5's failure, send NODE_FAILREP
    + NF_COMPLETEREPs to MGMD
 5  MGMD processes NODE_FAILREP + NF_COMPLETEREP, clear's node 5's cached address
 6  MGMD receives MGM protocol request for data node 5's address
 7  Data Node 5 is *still connected*, and so its address is looked up and cached
 8  Eventually data node 5 disconnect is processed
 9  Later data node 5 connects from a new address 'host B'
 10 MGMD received MGM protocol request for data node 5's address -
    returns cached 'hostA' address

The problematic events are :
  - Slow disconnect processing (3 .... 8)
  - MGM protocol request arrives during slow disconnect (6,7), after node
    failure signal handling (5)

This situation is avoided by adding a cached node address clear step
as part of connecting to a node (e.g. at step 9) so that any previously
cached address is discarded and a new address can be used.

Change-Id: I88841acd44e654ebbdb9ce4c013f26324df45d30