Post-push fix for Bug#35814266 ut_a and ut_ad do not report values of LHS and RHS used in comparison
While const auto & lhs = (LHS) extends the lifetime of the LHS if it is
a temporary, it will not extend the lifetime of temporaries which were
used to compute the LHS as subexpressions.

So for example, if LHS was s.substr(7).c_str(), then it will not work
as expected because the s.substr(7) which is part of this expression
is in itself returing a temporary, lifetime of which is not extended.

Or to give another example, if we have get_foo() which returns
std::optional<std::string>, and LHS is *get_foo(), then the * operator
will give you a reference to the result of get_foo() subexpression,
which will be destructed after lhs is computed.
Basically, only the lifetime of the outermost expression is extended,
so if that whole expression itself references some temporary result of
its subexpression, then it will crash.

This patch simplifies this whole issue by simply copying the LHS to a
local variable: const auto lhs{LHS};.
In many cases the copying will probably be elided anyway, so it's small
price to pay for avoiding a rather difficult to spot issue (for example
gcc 13.2.0 does not warn about it even with -Wall -Wpedantic. The clang
17.0.1 does. MSVC 19.37 doesn't, even with /Wall, but InteliSense is
flagging it as an issue in IDE).

Change-Id: Ifb51614e7e3772dfb2f9b588608e0cb3cef69170