WL#15540 Transform correlated scalar subqueries: allow column references as function arguments
WL#13520 "Transform correlated scalar subqueries" has a requirement
(#9), which states that the equality expression operands must be
simple column references, e.g. in the below transformation example
from WL#13520, t2.a and t1.a — the operands of the WHERE clause
equality predicate — are simple references. This WL lifts this
restriction and allows the column references to be arguments of
functions.

SELECT * FROM t1 WHERE (SELECT a FROM t2 WHERE t2.a=t1.a) > 0;

->

(with non-strict grouping and a built-in assert to signal any
cardinality error):

SELECT t1.a AS a
FROM t1
    JOIN
    (SELECT t2.a AS a,
            COUNT(0) AS cnt
     FROM t2
     WHERE (t2.a > 0)
     GROUP BY t2.a) derived
WHERE t1.a = derived.a AND
      reject_if(derived.cnt > 1)

The outer reference, t1.a, can already be embedded as a function
argument, but the non-outer (inner for short) column that we place in
the GROUP BY above, t2.a, cannot.

With this WL, we will also be able to transform queries like

SELECT * FROM t1 WHERE (SELECT func(t2.a) FROM t2 WHERE func(t2.a) = t1.a ) > 0;

->

SELECT t1.*
FROM t1
     JOIN
     ( SELECT func(t2.a) AS a,
              COUNT(0) AS cnt
       FROM t2
       WHERE (t2.a > 0)
       GROUP BY func(t2.a)) derived
WHERE derived.a = t1.a AND
      reject_if(derived.cnt > 1)

This transformation is needed by Heatwave/RAPID, which doesn't support
subqueries in the WHERE clause.

Limitations: due to a phase problem in the resolver, functional
dependencies are not always handled correctly: if the transformed
subquery has explicit grouping, functional dependency analysis may be
too pessimistic and give an error when the transform is used (but not
without). See worklog for details.  A work-around is to wrap the
offending column in ANY_VALUE.

Change-Id: I06fc0bbf8931f87ca964dc784329f4b2e24e048f