Bug#35710404 mysql server 8.1.0 failed at free_blob_buffers_and_reset
UPDATE HISTOGRAM ON <column> USING DATA <json> creates a temporary Field
object for <column> that is used internally to validate the bucket
values in the user-supplied histogram by calling Field::store().

It turns out that the creation of a Field in this context can have
unintended side-effects, such as incrementing the blob_fields counter on
the TABLE_SHARE in the case of TEXT columns. This causes the server to
fail when it is closing the TABLE object and freeing up blob memory
because it attempts to access a field that does not exist.

This patch changes the validation logic for user-defined histograms to
use the existing Field object associated with <column> to validate the
user input. This avoids unintended side-effects and modifications to
TABLE_SHARE outside of mutex-protection. Using the existing Field object
to validate data by calling Field::store() does not interfere with the
UPDATE HISTOGRAM ... USING DATA command as it does not otherwise make
use of the record buffer.

Change-Id: I1e1700765e17a7c43ba95b55e0f38c1777a12333