Bug#31969607: THE RESULT IS INCONSISTENT BETWEEN JOIN AND GROUP BY
Wrong results were returned if an equijoin condition referred to a
JSON expression and the chosen join strategy was a hash join.

The reason was that the join condition used string comparison
semantics or floating-point comparison semantics, depending on which
type the JSON expression was compared to. Both approaches are wrong,
as the values should be compared as JSON values.

Fixed by using Json_wrapper::make_hash_key() to create the hash values
when one or both of the sides of the hash join condition is a JSON
expression.

Change-Id: I02dc157a51c433a41225c25b60f8b3569593d0d1
(cherry picked from commit 2a0bb9a47ca6833bf1efd145f0f415c4eb05166a)