Bug#35227319 String histogram collation mismatch on comparison with REVERSE(1)
When using a default string histogram on a TEXT column we get an
assertion error due to a collation mismatch when we trigger a comparison
between the histograms bucket values with the String that is returned by
REVERSE(1).

The histogram checks whether the collation of the Item to compare
against matches its own collation before attempting to perform a
comparison, and later asserts that this holds on the actual String
objects that it compares.

The histogram comparison uses the String returned by Item::val_str(). In
the case of REVERSE(<input>) the String returned by val_str() inherits
the collation of its input, but this is not always consistent with the
collation of the Item_func_reverse that represents REVERSE() itself. In
particular, for REVERSE(1), the input "1" uses the latin1 collation,
while the Item_func_reverse uses the default utf8mb4 collation.
Therefore the initial check that the collations between the Item and the
histogram match passes, but the assertion fails when comparing the
actual String objects.

This patch updates the val_str() method of Item_func_reverse to ensure
that the collation of the returned string matches the collation of
Item_func_reverse itself. This is done by converting the collation of
the input to REVERSE() to match the collation of the Item_func_reverse,
if needed.

Change-Id: I9d7ccc80a7a6e638d39c1f792358cff6f47c2cb7