mysqldump: add line break before each database row
It's interesting to see that at the end of the row there was a comment
that read 'Always row break' but there actually was no 'row break'.
It appears as if the idea _was_ that there was a newline after each row
of data, increasing the readability of the dump files.

However I opted to put the newline character _before_ the line instead,
so also the first row of output is on a separate line from the INSERT
statement.

Fixes https://bugs.mysql.com/bug.php?id=65465