Bug#35483044 Hypergraph: Invalid row estimate for filter on 'Index range scan'
This commit improves row estimates for the following cases:

- field <=> NULL : If field is non-nullable, then the selectivity is 0.
  This fixes the error described in the bug report, where the selectivity
  of such a term was estimated to be 0.1.

- field=NULL : The selectivity is 0.

- field <=> other_expression : Use the same algorithm as for
  'field=other_expression', meaning that we use histograms or index
  statistics if applicable.

This commit also changes mysql-test/include/index_merge1.inc so that
the code for populating table t0 is pure SQL rather than perl.
This makes it easier to run part of the index_merge_... tests in
isolation, by copying the relevant SQL code from a .result file.

Change-Id: Ibba50f61a60edc0184293edb7317c5be6e082223