Bug#35813111: Use C++ standard library for more bit operations [my_bit, noclose]
After the previous patches for this bug, there is only a single
function left in my_bit.h/my_bit.cc, and that is my_reverse_bits().
Since this function is used in just one single file, there is not much
need to have a library for it. This patch removes my_bit.h and
my_bit.cc, and moves the my_reverse_bits() function to a static
function in lf_hash.cc instead.

Change-Id: I87c61fb5e3315ffc6473e1c6fe7dfdf178b3fbcd