Bug#35763950 rebootClusterFromCompleteOutage hangs if all instances are reachable
Problem
==================================

After a major outage of a Cluster, if all instances are
reachable dba.rebootClusterFromCompleteOutage() hangs and
fail rejoining instances.

Error is also reproducible with steps below:

- Create cluster with 3 members;
- Stop primary node;
- Stop secondary nodes from cluster;
- Start all members of cluster;
- execute: dba.rebootClusterFromCompleteOutage("cluster1", {force:true});

Analysis and Fix
==================================

After a careful analysis, what happens in such a scenario
is that the old secondaries retain old credentials to connect
to the nodes. This causes what we call an half connection,
in which, when the 2nd server tries to rejoin the group,
is able to contact with the primary, send an add_node,
but then the primary is not able to contact it back.

Because the add_node was approved and the configuration was
installed on the first node, it already considers two nodes
for quorums, but in fact, the new view was never installed
on the group, because the first node was not able to
contact the second node.

The solution, is to validate if, when we receive an add_node,
if we have bi-directional communication available to the
new node.

Change-Id: I25ca42b89ba411b48cf6786347f328fc656b5367