Bug#35673915 Ndb BI setup can deadlock with concurrent DDL
Processing of DDL in ha_ndbcluster is not allowed until the
MySQLD is participating in the 'schema distribution' mechanism
of the MySQL Cluster.

This is currently indicated by the presence or absence of
an NdbApi event operation object pointer in the table share.

However :
 - These pointers are set as part of schema distribution setup,
   before schema distribution is available locally
 - These pointers remain set for some time after schema
   distribution stops, e.g. due to disconnection or BI restart.

This can result in DDL being allowed to try to start
concurrently with BI startup which can result in
delays and deadlocks.

This is fixed by :
 1.  Having DDL operations check whether the Server-local part
     of schema distribution is setup (BI is live)
 2.  Fixing a state hazard in the BI state info where the
     ndb_binlog_is_ready bool was not cleared when the BI is
     shutting down/restarting

Change-Id: Ibbf08532a5d1befd9bd52a6611fa5683e19a5699